requirements.txt
kafka-python
cassandra-driver
pyspark